Fixed upstream slave crash bug http://bugs.mysql.com/bug.php?id=66890.
Issue was due to the thd->mem_root heap being released prematurely during the
processing of deferred events. 

In sql/log_event.cc:6060 User_var_log_event::do_apply_event(), a test is made
to determine if deferred events are being collected and processed. If true, a
call to this->set_deferred is made to set a flag indicating that 'this' is a
deferred event. This means that the thd->mem_root should not be freed on the
completion of User_var_log_event::do_apply_event().

The problem lies in the way User_var_log_event::set_deferred() is implemented.
The class User_var_log_event has an internal member 'bool deferred' that
should be set to true when User_var_log_event::set_deferred() is called. The
original code was setting 'deferred' to the state of another member
'char* val'. At the time User_var_log_event::set_deferred() is being called,
the 'val' member is NULL, therefore leaving the 'deferred' member at 0/false
instead of the true that was intended.

Changing User_var_log_event::set_deferred() in sql/log_event.h:2596 to be
'{ deferred= true; }' instead of '{ deferred= val; }' fixes the issue. This
is easily seen by running the slave with Valgrind before and after.
Valgrind reports 15434 errors of both 'Invalid read' and 'Invalid write'
before the fix and 0 errors and 0 leaks after the fix.