Bug #719386: Query_time is increasing for every query a stored procedure logs to the slow log
The problem was that the slow query time code did not handle the case of
individual statements in stored procedures correctly, i.e. when
log_slow_sp_statements=1 (the default). The query time was calculate as
the different between the current time and the start of the CALL
statement executing the corresponding SP. Which led to multiple
problems, in particular to the cumulative time from the CALL statement
being used for long_query_time checks and the Query_time values in the
slow query log.

Fixed by temporarily resetting start_time and start_utime values in THD
before executing individual SP statements and restoring them to their
original (i.e. the ones for the CALL statement) values so that
cumulative time for the CALL statement is still handled appropriately.