merge trunk, resolving Makefile conflict (we are actuall 5.5.28-29.1
-----

Have Percona-Server-server depend on Percona-Server-shared

Percona-Server-server contains files that are also present in the stock
package mysql-libs. If mysql-libs was installed before, installing
Percona-Server-server would make the installation fail because of these
conflicting files.

Adding a Conflicts: mysql-libs would have Percona-Server-server conflict
with Percona-Server-shared since the latter provides mysql-libs. Instead
have Percona-Server-server depend directly on Percona-Server-shared so
that if mysql-libs is installed before Percona-Server-shared will be selected
for replacing mysql-libs as desired.

-----

Have Percona-Server-server depend on Percona-Server-client

Percona-Server-server needs my_check_defaults(1) for the installation, which
was contained in this package. However, in the stock package this program
is included in the client package. Trying to install Percona-Server-server
side by side with the stock client package would result in a conflict for
this program.

The implemented solution is to move my_check_defaults to the
client package, however also we depend from the server directly to the client
package by name. This is to ensure my_print_defaults(1) to be installed,
however we also prevent future incompatibilites with the stock mysql package.

This works with the previous commit which requires the shared package.

-----

Include a mysqld-debug binary in the debian binaries

Along with the binary built with RelWithDebInfo, build a new binary
mysqld-debug built as a Debug binary.

Include a --nodebug parameter in the debian binary to optionally not
build it.

-----

Update versions to 5.5.28-29.1

-----

Bump versions to 5.5.28-29.2

-----

bump version number to -29.3

-----

Update versions to 5.5.28-29.3

-----

Use libmysqlclient.so.16 from Percona Server for shared-compat

The libmysqlclient.so.16 library included in upstream's shared-compat
package, which we use for building ours, contains the YaSSL implementation
of the Crypto libraries, making it subject to bug #1028240.

This patch alters the build system for the shared-compat RPM so it uses
our own build of libmysqlclient.so.16, but still using the other versions
for building the package.