Use libmysqlclient.so.16 from Percona Server for shared-compat
The libmysqlclient.so.16 library included in upstream's shared-compat
package, which we use for building ours, contains the YaSSL implementation
of the Crypto libraries, making it subject to bug #1028240.

This patch alters the build system for the shared-compat RPM so it uses
our own build of libmysqlclient.so.16, but still using the other versions
for building the package.