Bug #1070856: Temporary files created by binary log cache are not purged               after transaction commit
The problem was that the temporary file used to cache transactions for
the binary log was initialized once (when the in-memory buffer was
exceeded) and close (and thus, removed) only when closing the
connection. Which might lead to problems with large temporary files
being created and not removed in case of persistent connections and
large transactions (i.e. LOAD DATA INFILE).

Fixed by truncating the temporary file, if used for a binary log
transaction cache, when committing or rolling back a statement or a
transaction.