Bug #830286: rows_read is overflowed when changing db
The problem was that Rows_read was calculated in a way which lead to a
negative value being printed in the slow query log. Another problem was
that nobody knew what Rows_read is actually supposed to be and thus, how
it should be calculated. Which made solving the former problem rather
difficult.

Fixed by making Rows_read to be a synonym for Rows_examined in the slow
query log. It was like that for SELECT queries before the fix for bug
#716210 had broken it completely. And for other queries it didn't make
much sense anyway.

Rows_read will be removed in Percona Server 5.6.