Enable SSL support without YaSSL
In the fix for bug lp:1028240, --with yassl was removed from the RPM
builder script. Internal logic inside percona-server.spec led to having
the server built without SSL support at all (--without-ssl) instead of
having it built with the libcrypto functions.

This fix ensures that the server is built --with-ssl.