BP: https://blueprints.launchpad.net/percona-server/+spec/file-format-detection-in-innochecksum
In some cases it is useful to know the version and/or format which was
used to create an InnoDB data file. For example, importing data files
created with versions before 5.1.7 may result in server crashes, because
those versions did not initialize some fields in tablespace/page headers
correctly (see the incompatibility note in the innodb_expand_export docs
referring to bug #1000221 and bug #727704).

The goal of this BP is to extend innochecksum with an option to read
file format information from a given InnoDB data file. As only the first
page needs to be read to detect the format/version information, it can
also be used on a running server.

This patch introduces the new '-f' option to innobackupex.  When the
'-f' option is specified, innochecksum reads the file format information
from a given InnoDB data file by checking the page and tablespace flags
available in the first page, and then exit instead of scanning the
entire file As only the first page needs to be read to detect the
format/version information, it can also be used on a running server.