Fix bug 1100159 (Failing assertion: __builtin_expect(!((size) & ((size) - 1)), (1)) in file buf0buddy.ic line 137).
The problem is that upstream revision 0.15725.39 changed the way page
control block is allocated in buf_LRU_free_block() from the buddy to
heap allocator.  XtraDB has extra mutex release, reacquisition and
block state re-check.  If the state re-check fails, then the page
control block is free'ed, and XtraDB was not updated to use the heap
instead of buddy deallocator.