Fix bug 1083669 (Inconsistent bitmap write error handling).
The bitmap file create error, if occurred in
log_online_start_bitmap_file(), used to be fatal to the server.
Adjust by making it non-fatal and making this function to return
success flag instead, and propagate it up to the correct handling
point: on server startup this error continues to be fatal, but during
log tracking this signals log tracking thread to stop.

For bitmap page write errors potentially occuring in
log_online_write_bitmap_page(), propagate them up to instead of
silently ignoring them.

If the bitmap write error happens during the user request handling,
then signal the thread to quit the next time it wakes up.