Fix bug 1108613 (Failing assertion: log_bmp_sys->parse_buf_end == log_bmp_sys->parse_buf in log0online.c line 1038).
The log tracker thread is unaware of the situation when the oldest
untracked log records are overwritten by the new log data.  The thread
attempts to read and parse such records, which results in bad changed
page data in the best case, but is more likely to result in assertion
errors in the log parser.

Fixed by stopping the log tracker in this case.  Added a test to
percona_changed_page_bmp.