Fix
- bug 1105709 (InnoDB: Error: Write to file ./ib_modified_log_1_0.xdb
  failed at offset 0 0. when O_DIRECT is used in combination with
  innodb_track_changed_pages=1) and
- bug 1111144 (Newly created bitmap files should indiscriminately
  overwrite old ones).

The issue with bug 1105709 is that bitmap files were created with a
os_file_create(..., OS_DATA_FILE, ...) call, that makes it to be
treated as an InnoDB data file  for flushing purposes.  Thus,
innodb_flush_method=O_DIRECT made  the bitmap files to be created with
O_DIRECT too, and they were not ready for such operation nor it was
intended to.

The fix is to switch to os_file_create_simple API that makes
innodb_flush_method stop applying to the bitmap files.  For best OS
cache effects, indicate that we are not going to read the newly
written bitmap data by a posix_fadvise() call.

Add a testcase percona_changed_page_bmp_flush that tests the
combination of bitmap tracking with all the InnoDB flushing methods.

The issue with bug 1111144 is that the os_file_create() above
specified OS_FILE_OVERWRITE too, thus any new bitmap files would
silently overwrite any existing files with the same name.  The switch
to os_file_create_simple() made the file create fail in the case of
existing file.  A special case is a zero-sized existing files, which
are part of a regular operation, and they are deleted before creating
a new file.