Fix bug 1114612 (Failing assertion: n % srv_log_block_size == 0 in file os0file.c line 4269).
The assertion happens because InnoDB I/O and AIO routines assume that
all I/O is done in log block size blocks.  This works OK with the
default log block size of 512, but is likely to assert with large
innodb_log_block_size values.

Fix by asserting that all I/O is done in the minimum allowed log block
size blocks, that is, 512 bytes.

Add a testcase percona_log_block_size that tests this by setting the
log block to 4K and attempting to create a compressed InnoDB table
with 1K page size.