Fix bug 1111226 (Writing larger than 4GB bitmap file fails; bitmap write error causes heap corruption).
The first issue is incorrect higher 32 bits of write offset
calculation: the 64-bit offset was shifted to the left instead of the
right, fixed trivially.

The second issue is that if a bitmap page write fails (for example,
due to the condition above), then log_online_write_bitmap() exits
prematurely with a return code signalling error.  But, if at least one
bitmap page was written successfully, then the written bitmap tree
nodes will be doubly pointed to: from the bitmap tree itself and from
the free node list.  This will cause a heap corruption on the log
following thread shutdown (which happens immediatelly in the case of
bitmap write error).

Fixed by keeping iterating over the bitmap tree in case of error, but
skipping the actual write.  This way all the tree nodes are moved to
free list for the subsequent cleanup.

Add a write failure injection site and a new testcase
percona_changed_page_debug.