Fix bug 1089265 (Unstable rpl_percona_crash_resistant_rpl test).
The problem is that the crashes injected by the test might leave
slave's master.info corrupted, leaving it unable to start.  As the
purpose of the test is to test relay-log.info and not master.info
corruption recovery, restructure the test so that the slave I/O thread
is stopped by the time the slave SQL thread runs and crashes, so that
master.info file is closed at that time.