Bug #1081016: mysqldump --innodb-optimize-keys may produce invalid SQL               with explicitly named FK constraints
The problem was that mysqldump --innodb-optimize-keys produced invalid
SQL for cases when there was an explicitly named FK constrained which
implied a implicit secondary index with the same name. An attempt to
skip such a key definition from table schema and add it later with an
ALTER TABLE statement resulted in an "Incorrect index name" error,
because an index with the same name has already been created implicitly.

Fixed by detecting such cases in skip_secondary_keys() and restoring
the secondary to table schema in case there's an explicitly named FK
constraint with the same name.