[thread_pool_high_prio_tickets] Implementation of: 'thread_pool_high_prio_tickets' configuration option
https://blueprints.launchpad.net/percona-server/+spec/priority-scheduling-for-threadpool

Introduce the new ' thread_pool_high_prio_tickets' configuration option
to thread pool implementation with the following semantics:

This variable controls the high priority queue policy. Each new
connection is assigned this many tickets to enter the high priority
queue. Whenever a query has to be queued to be executed later because no
threads are available, the thread pool puts the connection into the high
priority queue if the following conditions apply:

1. The connection has an open transaction in the server.
2. The number of high priority tickets of this connection is non-zero.

If both the above conditions hold, the connection is put into the high
priority queue and its tickets value is decremented. Otherwise the
connection is put into the common queue with the initial tickets value
specified with this option.

Each time the thread pool looks for a new connection to process, it
first checks the high priority queue, and picks connections from the
common queue only when the high priority one is empty.

The idea is to minimize the number of open transactions in the
server. In many cases it is beneficial to give short-running
transactions a chance to commit faster and thus deallocate server
resources and locks without waiting in the same queue with other
connections that are about to start a new transaction, or those that
have run out of their high priority tickets.

With the default value of 0, all connections are always put into the
common queue, i.e. no priority scheduling is used as in the original
implementation in |MariaDB|. The higher is the value, the more chances
each transaction gets to enter the high priority queue and commit before
it is put in the common queue.

-----

Have the server package require on client package and conflict mysql-libs

For RPM, the server package wasn't requiring on the client package,
leading to installing the server and with it potentially some scripts
that depended on binaries contained in the client package.

Also, the shared package didn't Conflict: or Obsolete: mysql-libs,
leading to file-level conflicts when installing if mysql-libs was
previously installed. Have it Obsolete: mysql-libs the same way
Percona-Server-shared-55 deals with it.

-----

Add new man pages to the file list (RPM, dpkg)