Change the shared-compat RPM name to Percona-Server-shared-compat-51
Since the shared-compat package didn't have the -51 suffix, it shared the
name (but not the version) with the 5.5 shared-compat package, making it
impossible to install it from repositories. If installing both
Percona-Server-shared-51 and the shared-compat package was desired, they
would conflict.

Alter the name to Percona-Server-shared-compat-51 and remove the
libmysqlclient*.so.16* libraries from it, so it won't conflict with the
shared package. This package should be installed only if having the
older libraries is desired.