Change BUILD_DEBUG_BINARY to SKIP_DEBUG_BINARY in dpkg
The --nodebug parameter in the debian builder script used a
BUILD_DEBUG_BINARY variable in the debian/rules file which had the
opposite value. This caused problems with the default values -- when
no option was given, BUILD_DEBUG_BINARY had to be set and for --nodebug
it had to be unset.

Change BUILD_DEBUG_BINARY to SKIP_DEBUG_BINARY so they can have the same
default value.