Apply changes from revno 343 from 5.5
Port the changes introduced in revision 343 of the 5.5 tree. This
introduced a --nodebug parameter for the debian builder and enabled use
of the updated dh rules style (with override_ targets).

This had been null-merged as 234.1.109 in the 5.6 tree so check the log
for details.

This change is required for easy merging further changes to the rules
file.