[security_auth] Bug 1171941: CVE-2012-5615 security bug.
There is a difference it authentication handshake in cace when user
was found and when it wasn't.

This is a port of MDEV-3909: remote user enumeration.

Original commit message:
instead of returning Access denied on the incorrect user name,
emulate the complete failed logic procedure, possibly with
the change plugin packet.