[innodb_log_file_resize] Fix for bug #1172591
If server starts and log file size is changed the current lsn of new
created log file is aligned up the log block size but min_flushed_lsn
in data file remains untouched. So archived logs are started with one lsn
value but data files contain the other flushed lsn value. That causes
the error during archived logs applying.

This is the fix of that situation. After new log file creation the start
lsn of this log is written as (min|max)_flushed_lsn to data files.