Fix bug 1131949 (server won't start when flush_method=ALL_O_DIRECT and track_changed_pages=TRUE).
The issue is that bitmap tracking performs the log read I/Os following
the same configuration as the log subsystem itself.  That is, if
innodb_flush_method=ALL_O_DIRECT, then the log files have O_DIRECT set
on them, and the read buffer must be block-aligned, but it was not.

Fix by aligning the log tracker log reader buffer at
OS_FILE_LOG_BLOCK_SIZE.