Fix bug 1186690 (MySQL clients return bogus errno for host-not-found errors on Ubuntu 13.04) / http://bugs.mysql.com/bug.php?id=69379.
The issue is that gethostbyname_r() started returning underdocumented
values for host-not-found errors in the libc of Ubuntu 13.04.  There
are two such cases:

1) It may return -1 in *h_errno, which seems to correspond to
NETDB_INTERNAL, and there are vague suggestions that the real errno
value is then set in the global (thread-local) errno instead of
h_errno arg.

2) It may the struct hostent **result arg to NULL and still return
success (0) as its return value.

Attempt to fix by the following in my_gethostbyname_r():
1) check both result arg and return value to detemine whether the libc
call succeeded;
2) if -1 is found in h_errno, assume it to be NETDB_INTERNAL and copy
errno to h_errno.

The fix correctness is not fully proven, especially for 2), and it
also loses the fact that NETDB_INTERNAL has occurred, but this appears
to be good enough until the upstream fixes it.