Get datadir from my_print_defaults in RPM install (fix bug #1181753)
For the RPM installation, the datadir was hardcoded to /var/lib/mysql/,
so for users that have the datadir installed elsewhere, mysql_install_db
was being called with the wrong datadir, recreating the databases at that
location.

Port the relevant part for the fix for this upstream bug from
http://lists.mysql.com/commits/130338
to fix this.