Bug 1192354: accessing information_schema.partitions causes plans to change ha_innodb::info
Original title:
Bug 1192354: accessing information_schema.partitions causes plans to change ha_innodb::info stores statistics for keys in the table structure, which is shared for all the partitions of the table. When ha_partition::info is invoked, it calls ha_innodb::info for every partition, determines which has the maximum number of rows and invokes ha_innodb::info again for this partition. It sets partitioned table statistics to the statistics of the largest part. When i_s.partitions is invoked, it calls ha_innodb::info for every partition as well, but it doesn't perform the last call to set statistics for the whole table back. So, the table statistics is set to statistics of the last query, which is not the best choice in most cases. This patch implements behavior for i_s.partitions similar to behavior of ha_partition::info.

ha_innodb::info stores statistics for keys in the table structure, which is
shared for all the partitions of the table.
When ha_partition::info is invoked, it calls ha_innodb::info for every
partition, determines which has the maximum number of rows and invokes
ha_innodb::info again for this partition. It sets partitioned table
statistics to the statistics of the largest part.
When i_s.partitions is invoked, it calls ha_innodb::info for every partition
as well, but it doesn't perform the last call to set statistics for the
whole table back. So, the table statistics is set to statistics of the last
query, which is not the best choice in most cases.
This patch implements behavior for i_s.partitions similar to behavior of
ha_partition::info.

Fix bug 1261341 (Revert our fix for bug 1192354 in 5.6).

The fix is a straightforward Percona fix revert while preserving the
testcase.