Disable DEBUG_EXTNAME in binaries to fix bug #1179359
In 5.6, debug binaries named mysqld-debug are enabled by default even if
the build is a debug build (this is different than in 5.5, where for
debug builds they are named mysqld). This causes subtle differences in
testing for 5.6 which may be undesirable for binary builds that are
specific for debugging (--debug).

Disable DEBUG_EXTNAME, which is the option that causes the debug binary
to be named mysqld-debug, specifically for binary builds whem --debug
was supplied.