Fix generated cmake/install_layout.cmake in SRPMS when cross-compiling
When cross-compiling to i686 with build-rpm.sh with the --i686 switch,
build-rpm.sh would patch cmake/install_layout.cmake moving the lib64/ dir
to lib/, as it should be for the target 32 bit platform. The rationale
behind this is that the library directory name (lib64 or lib) is
determined by the CMAKE_SYSTEM_PROCESSOR cmake variable, and there's no
documented way of altering this variable (setting it with -D doesn't
work), so when cross-compiling the libraries would get installed to
lib64/ anyway.

The problem was that this was done prior to creating the source .tar.gz,
so the sources that went into the SRPM would be patched, potentially
creating problems if later this SRPM was used for creating 64bit binaries
(bug #1174300).

Fix this by moving the patching logic to percona-server.spec, where the
cmake/install_layout.cmake is patched after uncompressing from source.