Remove the logic that creates my.cnf in basedir from mysql_install_db
This is a fix for upstream bug http://bugs.mysql.com/bug.php?id=68416.

Currently mysql_install_db installs a my.cnf file in basedir, which for
rpm and dpkg is /usr, with some common options. Since we don't want a
my.cnf created anywhere, this patch removes this code from the script
altogether.

In order to facilitate future merges, in particular if in the future the
bug gets fixed upstream (possibly adding an option not to install the
file but that keeps the code), the code is commented out and not deleted.

======================================================

Fixed search path for libraries

======================================================

Reverted changes to install_layout.cmake and fixes around that change

======================================================

Updated debian changelog

======================================================

Defines the variable $cnfext in mysql_install_db Fixes warning Useless use of a constant in void context in mysql_install_db

Fixes warning Useless use of a constant in void context in mysql_install_db

======================================================

Bug #1382782 fix.

The return value of 'close' perl function is checked to detect error
on mysqld execution during creating and filling system tables from
mysql_install_db script.