Do not package the libmysqlclient.so.18 links in the library packages
Do not package this links since they are created by ldconfig(8) anyway
after installation. This will prevent conflicts between higher versions
of the libraries.

For debian, this also means adding a manual link since on install links
that are not pointing nowhere won't be created, and the fully versioned
libmysqlclient_r library is a symbolic link to the unversioned library,
instead of to the fully versioned.