Bug 1248046: SSL certificate revocation list-related MTR tests fail on some platforms with
Original title:
Bug 1248046: SSL certificate revocation list-related MTR tests fail on some platforms with -DWITH_SSL=system The crl name is calculated by OpenSSL as a hash of the issuer plus .r0, .r1 etc (which is kind of collision resolution). The issue is that hash function changed between different versions of OpenSSL. This specific test expect CRL to be stored in "fc725416.r0" for CentOS 5 and in "ab8a3803.r0" for CentOS 6. Test fails on CentOS 5 and other systems with old version of OpenSSL. One way to fix the test case is to make a copy of CRL with desired name.

The crl name is calculated by OpenSSL as a hash of the issuer
plus .r0, .r1 etc (which is kind of collision resolution). The
issue is that hash function changed between different versions of
OpenSSL. This specific test expect CRL to be stored in
"fc725416.r0" for CentOS 5 and in "ab8a3803.r0" for
CentOS 6. Test fails on CentOS 5 and other systems with old
version of OpenSSL. One way to fix the test case is to make a
copy of CRL with desired name.