Fix bug 1260035 (PURGE CHANGED_PAGE_BITMAPS BEFORE purges too many bitmap files).
PURGE CHANGED_PAGE_BITMAPS BEFORE <lsn> would remove bitmap files up
to the file containing the data for <lsn>.  If it is in the middle of
a bitmap file, the file is removed, destroying the bitmap data from
<lsn> up to the file end.

Fix by applying a modified patch by Andrew Gaul: in
log_online_purge_changed_page_bitmaps() check the next file in the
sequence after the current one to see if the end LSN falls into a
middle of the current file.  Make an exception for end LSN ==
IB_ULONGLONG_MAX, as there can be no next file then.  Setup the file
range up to the maximum LSN value instead of purge LSN value so that
we seen the next file after the purge LSN to make a decision whether
the one file before it should be deleted.

Add a test to percona_changed_page_bmp_requests.