Fix bug 1260259 (Crash on querying I_S.xtradb_rseg when innodb_rollback_segments/innodb_undo_logs lower then default).
The issue was that InnoDB in 5.6 replaced the rseg list data structure
with an array, whose elements are allowed to be NULL.  The array
iteration code in i_s_xtradb_rseg_fill() in xtradb_i_s.cc was not
expecting this.  Fixed trivially (keeping the invalid code formatting
in the function to minimize the patch).

Created a test xtradb_rseg_bug_1260259,