[innodb_partial_file_io] Fix bug 1262500 (ibd files contain bogus trailing zeros and non-page size file sizes): in t
Original title:
Fix bug 1262500 (ibd files contain bogus trailing zeros and non-page size file sizes): in the case pwrite() was interrupted, the retried write would not advance the buffer pointer, resulting in data corruption.  Also, the case of a partial write writing 0 bytes was not handled.

size file sizes): in the case pwrite() was interrupted, the retried
write would not advance the buffer pointer, resulting in data
corruption.  Also, the case of a partial write writing 0 bytes was not
handled.

Fix by applying a patch by Andrew Gaul, and extending to handle
pread() as well, except that a read would not be restarted if pread()
returned 0 bytes read, as that indicates an EOF instead of a partial
read.