Fix for bug 758788 - mysql process crashes after setting innodb_dict_size Fix for bug 1250018 - innodb_dict_size_limit tries to do LRU eviction of an index that is in use
dict_table_LRU_trim was missing some critical checks for existing locks on tables and indexes. Cleaned this up a bit and changed the algo to be not so aggressive when trying to find and evict tables from the dict cache.

With some of the new functionality, added some more debug assertions that should trip if the table is being evicted when it shouldn't as well as watching for an unexpectedly deleted table in some places.

Created very simple debug test case for 758788.