This is backport of fix for bug #915814 from https://code.launchpad.net/~akopytov/percona-server/bug915814-5.5 revision 228.
Revision 443 of this branch already contains fixes for this bug but not all of them. The
differences between them are the following:
1) 5.5 fix uses macros for counting buffer size for query in replication slave
event handler while 5.1 fix does not;
2) QueryStripComments::set() - copy the next symbol from query string to buffer
only if it's index does not exceed query length;
3) QueryStripComments_Backup::QueryStripComments_Backup() - copy current DB name
length(thd->db_length) to the end of QueryStripComments::query() buffer after
terminating NULL;
4) Query_cache::send_result_to_client() - the same as 3;
5) percona_bug856404 test case.

All items except 1 are backported.