[multiple_user_locks_kill] A post-merge fix for mulitple-user-level-locks per connection.
Behaviour of metadata locking has changed in 5.6 so that whenever
a wait on a lock is killed, ER_QUERY_INTERRUPTED error is pushed
immediately into the diagnostics stack.

The original behaviour of GET_LOCK() is to return NULL
in case of error - including such case as a killed query.

However, when there is an error on the diagnostics stack,
query execution is aborted, regardless of what a function
used in a query returns.

The rumour is MySQL team decided to change GET_LOCK() behaviour in 5.7,
so that it doesn't have to convert errors to NULL. Good decision.

But for the original patch, I must adapt the code to preserve
5.5 behaviour and silence ER_QUERY_INTERRUPTED error,
so that GET_LOCK() can return NULL. Ugly, but compatible.

This fixes failing func_misc and rpl_stm_000001.test.